+++ packages/dashboard/src/pages/Settings/components/AppearanceSetting.tsx
@@ -1,10 +1,9 @@
 import { MenuItem, type SelectChangeEvent } from '@mui/material'
+import { Appearance } from '@masknet/public-api'
 import { Services } from '../../../API.js'
 import { useAppearance } from '../../Personas/api.js'
-
 import SettingSelect from './SettingSelect.js'
 import { useDashboardI18N } from '../../../locales/index.js'
-import { Appearance } from '@masknet/public-api'
 
 function handleChange(event: SelectChangeEvent<Appearance>) {
     Services.Settings.setTheme(event.target.value as Appearance)
