+++ packages/mask/src/components/DataSource/usePluginHostPermission.ts
@@ -1,8 +1,8 @@
 import { useEffect } from 'react'
 import { useAsyncFn, useAsyncRetry } from 'react-use'
 import type { Plugin } from '@masknet/plugin-infra'
-import Services from '../../extension/service.js'
 import { MaskMessages } from '@masknet/shared-base'
+import Services from '../../extension/service.js'
 
 export function usePluginHostPermissionCheck(plugins: Plugin.Shared.Definition[]) {
     const plugins_ = plugins.filter((x) => x.enableRequirement.host_permissions?.length)
