+++ packages/mask/src/components/GuideStep/index.tsx
@@ -1,10 +1,10 @@
 import { cloneElement, useRef, useState, useLayoutEffect, type ReactElement } from 'react'
-import { useValueRef } from '@masknet/shared-base-ui'
 import { makeStyles, usePortalShadowRoot } from '@masknet/theme'
 import { Box, Modal, styled, Typography } from '@mui/material'
+import { useValueRef } from '@masknet/shared-base-ui'
+import { sayHelloShowed, userGuideFinished, userGuideStatus } from '@masknet/shared-base'
 import { activatedSocialNetworkUI } from '../../social-network/index.js'
 import { useI18N } from '../../utils/index.js'
-import { sayHelloShowed, userGuideFinished, userGuideStatus } from '@masknet/shared-base'
 
 const useStyles = makeStyles()((theme) => ({
     container: {
