+++ packages/mask/src/components/GuideStep/useSetupGuideStatus.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 import { useValueRef } from '@masknet/shared-base-ui'
-import { activatedSocialNetworkUI } from '../../social-network/index.js'
 import { currentSetupGuideStatus, type SetupGuideContext } from '@masknet/shared-base'
+import { activatedSocialNetworkUI } from '../../social-network/index.js'
 
 export function useSetupGuideStatus() {
     const context = useValueRef(currentSetupGuideStatus[activatedSocialNetworkUI.networkIdentifier])
