+++ packages/mask/src/social-network/defaults/inject/CommentBox.tsx
@@ -1,11 +1,11 @@
 import { memo, useCallback, useContext } from 'react'
 import { type PostInfo, usePostInfoDetails, PostInfoContext } from '@masknet/plugin-infra/content-script'
 import { type DOMProxy, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { makeStyles } from '@masknet/theme'
+import { MaskMessages } from '@masknet/shared-base'
 import { CommentBox, type CommentBoxProps } from '../../../components/InjectedComponents/CommentBox.js'
 import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
-import { makeStyles } from '@masknet/theme'
 import { startWatch } from '../../../utils/watcher.js'
-import { MaskMessages } from '@masknet/shared-base'
 
 const defaultOnPasteToCommentBox = async (
     encryptedComment: string,
