+++ packages/shared-base/src/Messages/CrossIsolationEvents.ts
@@ -20,7 +20,7 @@ import type {
     SmartPayDialogEvent,
     RenameWalletEvent,
     FollowLensDialogEvent,
-} from './Mask.js'
+} from './Events.js'
 
 /**
  * @deprecated
