+++ packages/shared-base/src/index.ts
@@ -2,8 +2,9 @@ export * from '@masknet/base'
 export * from './constant.js'
 export * from './Account/index.js'
 export * from './MIME/index.js'
-export * from './Messages/Mask.js'
-export * from './Messages/Shared.js'
+export * from './Messages/Events.js'
+export * from './Messages/CrossIsolationEvents.js'
+export * from './Messages/index.js'
 export * from './NameService/index.js'
 export * from './NextID/index.js'
 export * from './NextID/types.js'
@@ -32,4 +33,3 @@ export * from './test-helpers/index.js'
 export * from './unsafe-html/index.js'
 export * from './utils/index.js'
 export * from './legacy-settings/index.js'
-export * from './messages.js'
